Fixed Warnings. Added Weights to Gui Removed unused Code.

+++ Source/GameFiles/ClientSide/ClientObjectHandler.cpp
@@ -61,7 +61,7 @@ bool Client::AddNewPlayerObject(const std::vector<std::string>& msgArray, const 
 			this->zCreated = true;
 		}
 	}
-	MaloW::Debug("Player ID " + MaloW::convertNrToString(ID) +" Added");
+	MaloW::Debug("Player ID " + MaloW::convertNrToString((float)ID) +" Added");
 	//Creates a StaticMesh from the given Filename
 	iMesh* mesh = this->zEng->CreateStaticMesh(filename.c_str(), position);
 	mesh->SetQuaternion(rotation);
@@ -128,7 +128,7 @@ bool Client::AddNewAnimalObject(const std::vector<std::string>& msgArray, const 
 			MaloW::Debug("C: Unknown Message Was sent from server " + (*it) + " in AddNewAnimalObject");
 		}
 	}
-	MaloW::Debug("Animal Added ID: " + MaloW::convertNrToString(ID));
+	MaloW::Debug("Animal Added ID: " + MaloW::convertNrToString((float)ID));
 	//Creates a StaticMesh from the given Filename
 	iMesh* mesh = this->zEng->CreateStaticMesh(filename.c_str(), position);
 	mesh->SetQuaternion(rotation);
@@ -214,7 +214,7 @@ bool Client::AddNewStaticObject(const std::vector<std::string>& msgArray, const 
 			MaloW::Debug("C: Unknown Message Was sent from server " + (*it) + " in AddNewStaticObject");
 		}
 	}
-	MaloW::Debug("Static Object Added ID: " + MaloW::convertNrToString(ID) + " " + filename);
+	MaloW::Debug("Static Object Added ID: " + MaloW::convertNrToString((float)ID) + " " + filename);
 	//Creates a StaticMesh from the given Filename
 	iMesh* mesh = this->zEng->CreateStaticMesh(filename.c_str(), position);
 	mesh->SetQuaternion(rotation);
@@ -308,7 +308,7 @@ bool Client::AddNewDynamicObject(const std::vector<std::string>& msgArray, const
 			MaloW::Debug("C: Unknown Message Was sent from server " + (*it) + " in AddNewDynamicObject");
 		}
 	}
-	MaloW::Debug("Dynamic Object Added ID: " + MaloW::convertNrToString(ID) + " " + dynamicObject->GetName());
+	MaloW::Debug("Dynamic Object Added ID: " + MaloW::convertNrToString((float)ID) + " " + dynamicObject->GetName());
 	//Creates a StaticMesh from the given Filename
 	iMesh* mesh = this->zEng->CreateStaticMesh(filename.c_str(), position);
 	mesh->SetQuaternion(rotation);
